Added things so result is calculated well

+++ AHP/AHP.DAL/Entities/Criteria.cs
@@ -13,7 +13,7 @@ namespace AHP.DAL.Entities
         public int CriteriaId { get; set; }
         public int Order { get; set; }
         public string CriteriaName { get; set; }
-        public float Priority { get; set; }
+        public double Priority { get; set; }
         public DateTime DateCreated { get; set; }
         public DateTime DateUpdated { get; set; }
